# crates/wormhole-cli/tests/fixtures/local-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wormhole Local API",
    "description": "Control the per-user Wormhole daemon, manage exposures, inspect requests, and diagnose local providers. Operational routes require the bearer token stored in the private runtime directory.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:52731",
      "description": "Local Wormhole daemon"
    }
  ],
  "paths": {
...
    "/v1/endpoints": {
      "get": {
        "tags": [
          "Endpoints"
        ],
        "summary": "List active endpoints",
        "operationId": "endpoints",
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveEndpoint"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/endpoints/{id}": {
      "delete": {
        "tags": [
          "Endpoints"
        ],
        "summary": "Close or forget an endpoint",
        "operationId": "delete_endpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "forget",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedResponse"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/openapi.json": {
      "get": {
        "tags": [
          "Documentation"
        ],
        "summary": "Download the OpenAPI document",
        "operationId": "openapi_json",
        "responses": {
          "200": {
            "description": "OpenAPI"
          }
        },
        "security": [
          {}
        ]
      }
    },
...
    "/v1/remotes": {
      "get": {
        "tags": [
          "Remotes"
        ],
        "summary": "List configured relays",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoteView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Remotes"
        ],
        "summary": "Add or enroll a relay",
        "operationId": "add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteView"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/remotes/{name}": {
      "delete": {
        "tags": [
          "Remotes"
        ],
        "summary": "Remove a configured relay",
        "operationId": "remove",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "List captured requests",
        "operationId": "requests",
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapturedRequest"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Requests"
        ],
        "summary": "Clear captured requests",
        "operationId": "clear_requests",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests/{id}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Get a captured request",
        "operationId": "request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapturedRequest"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/v1/requests/{id}/replay": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Replay a captured request",
        "operationId": "replay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/v1/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List services",
        "operationId": "services",
        "parameters": [
          {
            "name": "watch",
            "in": "query",
            "description": "Wait for the next status change",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Create a service",
        "operationId": "create_service",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveEndpoint"
                  }
                }
              }
            }
          },
          "207": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveEndpoint"
                  }
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/services/{name}": {
      "delete": {
        "tags": [
          "Services"
        ],
        "summary": "Stop or forget a service",
        "operationId": "delete_service",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "forget",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedResponse"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/share": {
      "post": {
        "tags": [
          "Sharing"
        ],
        "summary": "Create an expiring share link",
        "operationId": "share",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActiveEndpoint": {
        "type": "object",
        "description": "A live or recently stopped exposure.",
        "required": [
          "id",
          "service",
          "driver",
          "urls",
          "status",
          "since"
        ],
        "properties": {
          "buffered_delivered": {
            "type": "integer",
            "format": "int64",
            "description": "Number of buffered webhooks delivered in this daemon lifetime.",
            "minimum": 0
          },
          "buffered_failed": {
            "type": "integer",
            "format": "int32",
            "description": "Number quarantined at the relay when last reported.",
            "minimum": 0
          },
          "buffered_pending": {
            "type": "integer",
            "format": "int32",
            "description": "Number waiting at the relay when last reported.",
            "minimum": 0
          },
          "driver": {
            "type": "string",
            "description": "Driver registry name."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Manager-local endpoint identifier."
          },
          "service": {
            "type": "string",
            "description": "Owning service name."
          },
          "since": {
            "type": "string",
            "format": "date-time",
            "description": "Time this status record was created."
          },
          "status": {
            "$ref": "#/components/schemas/EndpointStatus",
            "description": "Current lifecycle status."
          },
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Public URLs currently assigned."
          }
        }
      },
...
      "ClosedResponse": {
        "type": "object",
        "required": [
          "closed"
        ],
        "properties": {
          "closed": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}